попытка сделать ввод умножения без экранирования

+++ include/task.h
@@ -8,7 +8,7 @@ struct Task
     int value1;       // Первое число
     int value2;       // Второе число
     char operation;   // Операция (+, -, *, /, ^, !)
-    int result;       // Результат вычисления
+    double result;    // Результат вычисления
     int status;       // 0 = ok, -1 = деление на 0, 1 = неизвестная операция
 };
